+++ packages/jest-editor-support/src/types.js
@@ -12,6 +12,10 @@ export type Location = {
   line: number,
 };
 
+export type SpawnOptions = {
+  shell?: boolean,
+};
+
 import type {ChildProcess} from 'child_process';
 import type ProjectWorkspace from './project_workspace';
 
@@ -19,9 +23,11 @@ export type Options = {
   createProcess?: (
     workspace: ProjectWorkspace,
     args: Array<string>,
+    options?: SpawnOptions,
   ) => ChildProcess,
   testNamePattern?: string,
   testFileNamePattern?: string,
+  shell?: boolean,
 };
 
 /**